Rollup plugin: support app usage
The rollup plugin as currently written only really works for library development, where you want to emit separate CSS files that are imported from your published JS files.

But we'll also want to use it in apps, where those apps are building with Vite (or directly with Rollup).

This PR simplifies the plugin so it just worries about emitting CSS and doesn't care exactly how that CSS gets further processed.

To make this continue to work in v2 addons, I've [adjusted the implementaiton of the keepAssets plugin](https://github.com/embroider-build/embroider/pull/2156) so that it will compose nicely with a plugin like this one that emits CSS. So you'll want to update `@embroider/addon-dev` to a version that includes that PR.

In Vite, it just works with no further configuration since vite handles CSS imports out-of-the-box.